Explicitly cast NCONTINENTS to int to silence an overflow warning.
Former-commit-id: 1096ac7af442bfa36e6fc046f756d5860c8db831

+++ usr.sbin/tzsetup/tzsetup.c
@@ -473,7 +473,7 @@ begin_define
 define|#
 directive|define
 name|NCONTINENTS
-value|((sizeof continents)/(sizeof continents[0]))
+value|(int)((sizeof continents)/(sizeof continents[0]))
 end_define
 
 begin_define
